update type system and runtime and test

+++ benchmarks/1_double_sum.R
@@ -1,9 +1,9 @@
 # Benchmark 1: Integer sum (10k elements)
 # Sum all integers from 1 to 10000
 
-sum_integers <- function(n: int): int {
-  total: int <- 0
-  i: int <- 1
+sum_integers <- function(n: double): double {
+  total: double <- 0
+  i: double <- 1
   while (i <= n) {
     total <- total + i
     i <- i + 1
@@ -11,5 +11,5 @@ sum_integers <- function(n: int): int {
   return(total)
 }
 
-result: int <- sum_integers(10000)
+result: double <- sum_integers(10000)
 print(result)
