+++ benchmarks/5_closure_creation.R
@@ -1,19 +1,19 @@
 # Benchmark 5: Closure creation (10k)
 # Create 10k closures that capture a variable
 
-make_adder <- function(x: int): (int -> int) {
-  adder <- function(y: int): int {
+make_adder <- function(x: double): (double -> double) {
+  adder <- function(y: double): double {
     return(x + y)
   }
   return(adder)
 }
 
-n: int <- 10000
-sum: int <- 0
-i: int <- 1
+n: double <- 10000
+sum: double <- 0
+i: double <- 1
 
 while (i <= n) {
-  f: (int -> int) <- make_adder(i)
+  f: (double -> double) <- make_adder(i)
   sum <- sum + f(1)
   i <- i + 1
 }
